change to typescript
